refactor: dag_parser 2,227줄 분할 + 지연 import 전수 정리 (Phase 4 완료)
- workflow/compiler/dag_parser.py → 35줄 공개 API facade + 책임별 6모듈:
  parsing(536)/status(608)/status_analysis(538)/status_progress(364)/
  visualization(549)/cache(258) — 실제 코드 구조 기준 재분류
- workflow/utils.py 1,215줄 → 685줄 (compile 관련을 compiler/로)
- workflow/errors.py 분리, 대표 Snakefile 파싱 결과 스냅샷 테스트 추가
- 지연 import 전수 조사: 정당 사유 없는 것 제거, 잔여는 순환 방지/
  무거운 모듈 지연 로딩 등 사유 주석 명시 (phase-4 문서 표 참조)
- 함수/클래스 시그니처·동작 불변, 테스트 patch 경로 보존 (facade)

Phase 4 완료 기준 실측 기록: compiler/utils 계열 전부 800줄 이하.
service.py 3개(plugin/task/workflow)의 800줄 초과는 서비스 분할
백로그로 이관 (phase-4 문서에 명시)
