+++ backend/app/task/service.py
@@ -132,6 +132,9 @@ def get_task_monitoring(*, db: Session, current_user: models.User) -> List[TaskM
 def revoke_task(*, db: Session, current_user: models.User, task_id: str) -> dict:
     """Revoke a Celery task and clean up its containers + results folder."""
     try:
+        # 지연 import 유지 (사유): 테스트 patch 호환(`app.task.service.get_celery_app`가
+        # 아닌 호출부 재바인딩 방식) + worker가 task.service를 import 해도 web 계층
+        # (app.main)이 로드되지 않도록 하기 위함 (import-linter no-web-in-worker).
         from app.main import get_celery_app
         celery = get_celery_app()
